Sort out constructor parameters
... to match the order from the parent storage class.

Also:
  * Make `DsIdentifiers` final.
  * Move `EntityColumnPredicate` to the package level.
  * Ignore Datastore credentials file by mask.
