+++ .gitignore
@@ -27,7 +27,7 @@ gcs-auth-key.json
 # They are stored in the encrypted file and are decrypted on the CI server only.
 
 # This file extracted form credentials.tar
-spine-dev.json
+spine-dev*.json
 /datastore/src/test/resources/spine-dev-62685282c0b9.json
 
 package-lock.json
